+++ rust/patchable/src/patch_mail.rs
@@ -80,6 +80,8 @@ pub fn mailsplit(repo: &Repository, patch_file: &Path) -> Result<impl Iterator<I
     let mailsplit = suspend_tracing_indicatif(|| {
         raw_git_cmd(repo)
             .arg("mailsplit")
+            // From <commit> is ignored anyway, so there's no point requiring it
+            .arg("-b")
             // mailsplit doesn't accept split arguments ("-o dir")
             .arg({
                 let mut output_arg = OsString::from("-o");