+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
   util-linux-core contains a basic set of Linux utilities, including the
   command logger which allows to enter messages into the system log.
 - vector: Add version 0.43.1 ([#980]).
+- druid: Add version 30.0.1 and 31.0.1 ([#984])
 - opa: Add version 1.0.1 ([#981], [#1000]).
 - statsd-exporter: Bump version to 0.28.0 ([#982]).
 - git-sync: Bump version to 4.4.0 ([#990]).
@@ -36,7 +37,7 @@ All notable changes to this project will be documented in this file.
 - kafka: Bump 3.8.0 to 3.8.1 ([#995]).
 - Update registry references to oci ([#989]).
 - trino-storage-connector: Move the build out of trino/ for easier patching ([#996]).
-- druid 26.0.0: Migrate to patchable ([#1003]).
+- BREAKING: druid: Bump opa-authorizer to 0.6.0 for all versions ([#984])
 
 ### Removed
 
@@ -47,6 +48,7 @@ All notable changes to this project will be documented in this file.
 - trino-cli: Remove version 469 ([#999]).
 - trino-storage-connector: Remove version 469 ([#999]).
 - nifi: Remove 2.0.0 ([#1006]).
+- druid: Remove 26.0.0 ([#984])
 
 ### Fixed
 
@@ -66,9 +68,10 @@ All notable changes to this project will be documented in this file.
 [#980]: https://github.com/stackabletech/docker-images/pull/980
 [#981]: https://github.com/stackabletech/docker-images/pull/981
 [#982]: https://github.com/stackabletech/docker-images/pull/982
+[#984]: https://github.com/stackabletech/docker-images/pull/984
+[#988]: https://github.com/stackabletech/docker-images/pull/988
 [#989]: https://github.com/stackabletech/docker-images/pull/989
 [#990]: https://github.com/stackabletech/docker-images/pull/990
-[#988]: https://github.com/stackabletech/docker-images/pull/988
 [#991]: https://github.com/stackabletech/docker-images/pull/991
 [#992]: https://github.com/stackabletech/docker-images/pull/992
 [#993]: https://github.com/stackabletech/docker-images/pull/993